requirements.txt
aio-pika~=9.5.8
aiormq~=6.9.2
fastapi~=0.121.3
SQLAlchemy~=2.0.44
pydantic~=2.12.4
python-dotenv~=1.2.1
starlette~=0.50.0
httpx~=0.28.1
pytest~=9.0.1
pytest-asyncio~=1.3.0
alembic~=1.17.2
pydantic-settings~=2.12.0
asyncpg~=0.31.0
uvicorn==0.30.5
ruff==0.6.0